z80 opcodes cycles were synchronized with ula refresh rate

+++ src/render/render.h
@@ -2,7 +2,7 @@
 #define _RENDER_H_
 #include <stdint.h>
 
-void render_init(unsigned long* display_buffer, size_t buffer_size_x, 
+void render_init(uint32_t* display_buffer, size_t buffer_size_x, 
 	size_t buffer_size_y, size_t display_size_x, size_t display_size_y,
 	size_t window_size_x, size_t window_size_y);
 void render_end();
